- Finished working on the node_spec feature - Polished the user accounts module - Next up we will focus on page parent-child relation.

+++ oy/contrib/admin/wrappers.py
@@ -7,7 +7,7 @@ from flask_admin.helpers import get_redirect_target
 from flask_admin.model.template import (
     DeleteRowAction,
     LinkRowAction,
-    EndpointLinkRowAction,
+    EndpointLinkRowAction
 )
 from flask_admin.contrib.sqla import ModelView, form
 from flask_security import current_user
@@ -16,19 +16,31 @@ from oy.babel import gettext, lazy_gettext
 
 
 class AuthenticationViewMixin:
+    """Basic permissions for oy views."""
+
     def is_accessible(self):
         if not current_user.is_active or not current_user.is_authenticated:
             return False
-        if current_user.has_role("admin"):
+        if current_user.has_role("staff"):
             return True
         return False
 
-    def _handle_view(self, name, **kwargs):
-        if not self.is_accessible():
-            if current_user.is_authenticated:
-                abort(403)
-            else:
-                return redirect(url_for("security.login", next=request.url))
+    def inaccessible_callback(self, name, **kwargs):
+        if current_user.is_authenticated:
+            abort(403)
+        else:
+            return redirect(url_for("security.login", next=request.url))
+
+
+class OyDeleteRowAction(EndpointLinkRowAction):
+
+    def __init__(self):
+        super().__init__(
+            "fa fa-trash",
+            ".delete_confirm",
+            title=lazy_gettext("Delete record"),
+            id_arg="pk"
+        )
 
 
 class OyModelFormConverter(form.AdminModelConverter):
@@ -46,12 +58,13 @@ class OyModelFormConverter(form.AdminModelConverter):
 
 
 class OyModelView(AuthenticationViewMixin, ModelView):
-    """Automatic authentication and some extras"""
+    """The base view class for all oy model views."""
 
     model_form_converter = OyModelFormConverter
 
     def __init__(self, *args, **kwargs):
         self.verbose_name = kwargs.pop("verbose_name", kwargs.get("name"))
+        self.show_in_menu = kwargs.pop("show_in_menu", True)
         super().__init__(*args, **kwargs)
 
     def get_column_name(self, field):
@@ -66,14 +79,7 @@ class OyModelView(AuthenticationViewMixin, ModelView):
         actions = super().get_list_row_actions()
         if self.can_delete:
             actions = [a for a in actions if not isinstance(a, DeleteRowAction)]
-            actions.append(
-                EndpointLinkRowAction(
-                    "fa fa-trash",
-                    ".delete_confirm",
-                    title=gettext("Delete"),
-                    id_arg="pk",
-                )
-            )
+            actions.append(OyDeleteRowAction())
         if hasattr(self, "get_preview_url"):
             actions.append(
                 LinkRowAction(
@@ -84,11 +90,14 @@ class OyModelView(AuthenticationViewMixin, ModelView):
             )
         return actions
 
-    @expose("/delete-confirm/<string:pk>")
+    @expose("/delete-confirm/<int:pk>")
     def delete_confirm(self, pk):
-        obj = self.get_one(pk)
+        if not self.can_delete:
+            abort(404)
+        obj = self.get_one(str(pk))
         if obj is not None:
-            return_url = get_redirect_target() or url_for(".index_view")
+            endpoint = ".index_view" if self.show_in_menu else "admin.index"
+            return_url = get_redirect_target() or url_for(endpoint)
             return self.render(
                 "admin/oy/delete_confirm.html", pk=pk, return_url=return_url
             )
@@ -97,12 +106,15 @@ class OyModelView(AuthenticationViewMixin, ModelView):
 
     @expose("/do-delete/<string:pk>")
     def delete_executor(self, pk):
+        if not self.can_delete:
+            abort(404)
         obj = self.get_one(pk)
         if obj is not None:
             db.session.delete(obj)
             db.session.commit()
             flash("Record deleted successfully.")
-            return redirect(get_redirect_target() or url_for(".index_view"))
+            endpoint = ".index_view" if self.show_in_menu else "admin.index"
+            return redirect(get_redirect_target() or url_for(endpoint))
         else:
             abort(404)
 
@@ -114,7 +126,11 @@ class OyIndexView(AuthenticationViewMixin, AdminIndexView):
 
 
 class OyBaseView(BaseView, AuthenticationViewMixin):
-    pass
+
+    def __init__(self, *args, **kwargs):
+        self.verbose_name = kwargs.pop("verbose_name", kwargs.get("name"))
+        self.show_in_menu = kwargs.pop("show_in_menu", True)
+        super().__init__(*args, **kwargs)
 
 
 def admin_required(f):
